Refactors Navigation menu files
