+++ src/components/Navigation/Desktop/index.js
@@ -1,0 +1,1 @@
+export { default } from './NavigationDesktop'
