+++ src/components/Navigation/Mobile/NavigationMobile.js
@@ -5,9 +5,10 @@ import faBars from '@fortawesome/fontawesome-free-solid/faBars'
 import faTimes from '@fortawesome/fontawesome-free-solid/faTimes'
 import { Link } from 'react-scroll'
 
-import { media } from '../../styles'
+import { media } from '../../../styles'
 
-import { LOCATIONS } from './constants'
+import { LOCATIONS } from '../constants'
+import MobileItem from './MobileItem'
 
 const NavigationMobileStyled = styled.button`
   float: right;
@@ -55,10 +56,6 @@ const MenuContentStyled = styled.ul`
   padding: 20px;
 `
 
-const MenuItemStyled = styled.ul`
-  margin: 26px 0;
-`
-
 class NavigationMobile extends Component {
   state = {
     menuOpen: false,
@@ -85,18 +82,9 @@ class NavigationMobile extends Component {
       </MenuHeaderStyled>
       <MenuContentStyled>
         {LOCATIONS.map(({ target, text }) => (
-          <Link
-            key={target}
-            to={target}
-            smooth
-            duration={500}
-            offset={-60}
-            onClick={this.toggleMenu}
-          >
-            <MenuItemStyled key={target} target={target}>
-              {text}
-            </MenuItemStyled>
-          </Link>
+          <MobileItem key={target} target={target} onClick={this.toggleMenu}>
+            {text}
+          </MobileItem>
         ))}
       </MenuContentStyled>
     </OpenMenuStyled>
